Break resultFields validation into separate module.
